Add compare_bytes_test.go (#40)
