+++ struct_serializer.go
@@ -123,33 +123,3 @@ func (structSerializer *structSerializer) scalarOrNullSplitParts(
 
 	return structSerializer.baseSerializer.scalarOrNullSplitParts(ionType, isNull, bytes)
 }
-
-func compareBytes(bytes1, bytes2 []byte) int {
-	for i := 0; i < len(bytes1) && i < len(bytes2); i++ {
-		byte1 := bytes1[i]
-		byte2 := bytes2[i]
-		if byte1 != byte2 {
-			return int(byte1 - byte2)
-		}
-	}
-
-	return len(bytes1) - len(bytes2)
-}
-
-// sortableBytes implements the sort.Interface so we can sort fieldHashes in stepOut()
-type sortableBytes [][]byte
-
-func (sb sortableBytes) Len() int {
-	return len(sb)
-}
-
-func (sb sortableBytes) Less(i, j int) bool {
-	bytes1 := sb[i]
-	bytes2 := sb[j]
-
-	return compareBytes(bytes1, bytes2) < 0
-}
-
-func (sb sortableBytes) Swap(i, j int) {
-	sb[i], sb[j] = sb[j], sb[i]
-}
